Extension + UITableViewController, same settings for tab/nav/tableVCs, ApplyForServiceTableViewController typesOfServices fix
