+++ LookLabApp/ViewControllers/AccountViewController.swift
@@ -32,6 +32,7 @@ final class AccountViewController: UIViewController {
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
+        
         tabBarController?.tabBar.isHidden = false
         updateAppointments()
     }
