+++ LookLabApp/ViewControllers/MasterListViewController.swift
@@ -8,6 +8,7 @@ final class MasterListViewController: UITableViewController {
     //MARK: - View life Cycles
     override func viewDidLoad() {
         super.viewDidLoad()
+        
         view.backgroundColor = .systemBrown
         enableBarSettings()
         setTablewViewBackgroundToGradient()
@@ -93,32 +94,6 @@ final class MasterListViewController: UITableViewController {
             let master = filterMasterOfType(masters: masters, section: 2)
             selectMasterVC.master = master[index.row]
         }
-    }
-    
-    //MARK: - Background & Bar Settings
-    private func setTablewViewBackgroundToGradient() {
-        let gradientLayer = CAGradientLayer()
-        let gradientBackgroundColors = [UIColor.clear.cgColor,
-                                         UIColor.black.cgColor]
-        
-        gradientLayer.colors = gradientBackgroundColors
-        gradientLayer.locations = [0.0, 1.0]
-        
-        gradientLayer.frame = tableView.bounds
-        let backgroundView = UIView(frame: tableView.bounds)
-        backgroundView.layer.insertSublayer(gradientLayer, at: 0)
-        tableView.backgroundView = backgroundView
-    }
-    
-    private func enableBarSettings() {
-        navigationController?.navigationBar.barTintColor = .systemBrown
-        
-        let appearance = UITabBarAppearance()
-        appearance.configureWithOpaqueBackground()
-        appearance.backgroundColor = .black
-        
-        tabBarController?.tabBar.standardAppearance = appearance
-        tabBarController?.tabBar.scrollEdgeAppearance = tabBarController?.tabBar.standardAppearance
     }
     
    // MARK: - Private func
